# openapi/openapi.bundle.yaml
openapi: 3.1.0
info:
  title: Python Project Inspector
  version: 0.1.0
paths:
...
  /api/graph:
    get:
      description: Return graph nodes and edges at one commit.
      operationId: graph_api_graph_get
      parameters:
        - in: query
          name: commit
          required: false
          schema:
            anyOf:
              - type: string
              - type: 'null'
            title: Commit
        - in: query
          name: include_zero_score
          required: false
          schema:
            default: false
            title: Include Zero Score
            type: boolean
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GraphResponse'
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Graph
...
  /api/v1/entities:
    get:
      operationId: listEntitiesV1
      parameters:
        - in: query
          name: entityKindId
          required: true
          schema:
            minLength: 1
            title: Entitykindid
            type: string
        - in: query
          name: commitId
          required: false
          schema:
            anyOf:
              - type: string
              - type: 'null'
            title: Commitid
        - in: query
          name: limit
          required: false
          schema:
            default: 5000
            maximum: 10000
            minimum: 1
            title: Limit
            type: integer
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ListEntitiesV1Response'
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: List entities of one kind at a commit
      tags:
        - entities
  /api/v1/graph:
    get:
      operationId: getGraphV1
      parameters:
        - in: query
          name: lensId
          required: true
          schema:
            minLength: 1
            title: Lensid
            type: string
        - in: query
          name: commitId
          required: false
          schema:
            anyOf:
              - type: string
              - type: 'null'
            title: Commitid
        - in: query
          name: includeZeroWeight
          required: false
          schema:
            default: false
            title: Includezeroweight
            type: boolean
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GraphV1Response'
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Generic graph projection
      tags:
        - graph
...
  /api/v1/tables/{tableId}:
    get:
      operationId: getTableV1
      parameters:
        - in: path
          name: tableId
          required: true
          schema:
            title: Tableid
            type: string
        - in: query
          name: commitId
          required: false
          schema:
            anyOf:
              - type: string
              - type: 'null'
            title: Commitid
        - in: query
          name: parentEntityId
          required: false
          schema:
            anyOf:
              - type: string
              - type: 'null'
            title: Parententityid
        - in: query
          name: limit
          required: false
          schema:
            default: 500
            maximum: 5000
            minimum: 1
            title: Limit
            type: integer
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TableV1Response'
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      summary: Generic table projection
      tags:
        - tables
...
components:
  schemas:
...
    EdgeResponse:
      properties:
        breakdown:
          anyOf:
            - additionalProperties:
                type: integer
              type: object
            - type: 'null'
          title: Breakdown
        commit_hash:
          title: Commit Hash
          type: string
        kind_occurrence_count:
          default: 0
          title: Kind Occurrence Count
          type: integer
        kinds:
          additionalProperties:
            type: integer
          title: Kinds
          type: object
        score:
          title: Score
          type: integer
        source:
          title: Source
          type: string
        target:
          title: Target
          type: string
      required:
        - source
        - target
        - score
        - kinds
        - commit_hash
      title: EdgeResponse
      type: object
...
    EntityRefV1:
      properties:
        id:
          title: Id
          type: string
        kind:
          title: Kind
          type: string
        label:
          title: Label
          type: string
        path:
          anyOf:
            - type: string
            - type: 'null'
          title: Path
        pluginId:
          anyOf:
            - type: string
            - type: 'null'
          title: Pluginid
      required:
        - id
        - kind
        - label
      title: EntityRefV1
      type: object
    EntityTargetV1:
      properties:
        id:
          title: Id
          type: string
        kind:
          title: Kind
          type: string
        label:
          title: Label
          type: string
        path:
          anyOf:
            - type: string
            - type: 'null'
          title: Path
        pluginId:
          anyOf:
            - type: string
            - type: 'null'
          title: Pluginid
        reason:
          anyOf:
            - type: string
            - type: 'null'
          title: Reason
        selectable:
          default: true
          title: Selectable
          type: boolean
      required:
        - id
        - kind
        - label
      title: EntityTargetV1
      type: object
...
    GraphEdgeV1:
      properties:
        attributes:
          additionalProperties: true
          title: Attributes
          type: object
        contributions:
          items:
            additionalProperties: true
            type: object
          title: Contributions
          type: array
        id:
          title: Id
          type: string
        metrics:
          items:
            $ref: '#/components/schemas/MetricValueV1'
          title: Metrics
          type: array
        relationTypeId:
          title: Relationtypeid
          type: string
        source:
          $ref: '#/components/schemas/EntityRefV1'
        target:
          $ref: '#/components/schemas/EntityRefV1'
      required:
        - id
        - source
        - target
        - relationTypeId
      title: GraphEdgeV1
      type: object
...
    GraphNodeResponse:
      properties:
        line_counts:
          additionalProperties:
            type: integer
          title: Line Counts
          type: object
        metrics:
          additionalProperties:
            type: number
          title: Metrics
          type: object
        module_name:
          title: Module Name
          type: string
        total_lines:
          title: Total Lines
          type: integer
      required:
        - module_name
        - total_lines
        - metrics
        - line_counts
      title: GraphNodeResponse
      type: object
    GraphNodeV1:
      properties:
        attributes:
          additionalProperties: true
          title: Attributes
          type: object
        distributions:
          items:
            additionalProperties: true
            type: object
          title: Distributions
          type: array
        entity:
          $ref: '#/components/schemas/EntityRefV1'
        lineCounts:
          additionalProperties:
            type: integer
          title: Linecounts
          type: object
        metrics:
          items:
            $ref: '#/components/schemas/MetricValueV1'
          title: Metrics
          type: array
      required:
        - entity
      title: GraphNodeV1
      type: object
    GraphResponse:
      properties:
        commit_hash:
          title: Commit Hash
          type: string
        edges:
          items:
            $ref: '#/components/schemas/EdgeResponse'
          title: Edges
          type: array
        nodes:
          items:
            $ref: '#/components/schemas/GraphNodeResponse'
          title: Nodes
          type: array
      required:
        - commit_hash
        - nodes
        - edges
      title: GraphResponse
      type: object
    GraphV1Response:
      properties:
        commitId:
          title: Commitid
          type: string
        edges:
          items:
            $ref: '#/components/schemas/GraphEdgeV1'
          title: Edges
          type: array
        lensId:
          title: Lensid
          type: string
        metrics:
          items:
            $ref: '#/components/schemas/MetricDefV1'
          title: Metrics
          type: array
        nodes:
          items:
            $ref: '#/components/schemas/GraphNodeV1'
          title: Nodes
          type: array
        relationTypes:
          items:
            $ref: '#/components/schemas/RelationTypeV1'
          title: Relationtypes
          type: array
      required:
        - commitId
        - lensId
      title: GraphV1Response
      type: object
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          title: Detail
          type: array
      title: HTTPValidationError
      type: object
...
    ListEntitiesV1Response:
      properties:
        commitId:
          anyOf:
            - type: string
            - type: 'null'
          title: Commitid
        entityKindId:
          title: Entitykindid
          type: string
        items:
          items:
            $ref: '#/components/schemas/EntityTargetV1'
          title: Items
          type: array
      required:
        - entityKindId
      title: ListEntitiesV1Response
      type: object
...
    MetricDefV1:
      properties:
        defaultAggregation:
          default: mean
          title: Defaultaggregation
          type: string
        entityKinds:
          items:
            type: string
          title: Entitykinds
          type: array
        format:
          $ref: '#/components/schemas/MetricFormatV1'
        higherIsWorse:
          default: false
          title: Higherisworse
          type: boolean
        id:
          title: Id
          type: string
        label:
          title: Label
          type: string
        scope:
          title: Scope
          type: string
        supportedAggregations:
          items:
            type: string
          title: Supportedaggregations
          type: array
        supportedViews:
          items:
            type: string
          title: Supportedviews
          type: array
        valueType:
          title: Valuetype
          type: string
      required:
        - id
        - label
        - valueType
        - scope
      title: MetricDefV1
      type: object
    MetricFormatV1:
      properties:
        kind:
          default: integer
          title: Kind
          type: string
      title: MetricFormatV1
      type: object
...
    MetricValueV1:
      properties:
        aggregation:
          anyOf:
            - type: string
            - type: 'null'
          title: Aggregation
        metricId:
          title: Metricid
          type: string
        value:
          anyOf:
            - type: number
            - type: integer
            - type: string
            - type: boolean
            - type: 'null'
          title: Value
      required:
        - metricId
      title: MetricValueV1
      type: object
...
    RelationTypeV1:
      properties:
        defaultVisible:
          default: false
          title: Defaultvisible
          type: boolean
        group:
          anyOf:
            - type: string
            - type: 'null'
          title: Group
        id:
          title: Id
          type: string
        label:
          title: Label
          type: string
        pluginId:
          anyOf:
            - type: string
            - type: 'null'
          title: Pluginid
      required:
        - id
        - label
      title: RelationTypeV1
      type: object
...
    TableColumnV1:
      properties:
        align:
          anyOf:
            - type: string
            - type: 'null'
          title: Align
        format:
          anyOf:
            - additionalProperties: true
              type: object
            - type: 'null'
          title: Format
        id:
          title: Id
          type: string
        label:
          title: Label
          type: string
        metricId:
          anyOf:
            - type: string
            - type: 'null'
          title: Metricid
        sortable:
          default: true
          title: Sortable
          type: boolean
        valueType:
          default: string
          title: Valuetype
          type: string
        visibleByDefault:
          default: true
          title: Visiblebydefault
          type: boolean
        width:
          anyOf:
            - type: integer
            - type: 'null'
          title: Width
      required:
        - id
        - label
      title: TableColumnV1
      type: object
...
    TableRowActionV1:
      properties:
        id:
          title: Id
          type: string
        kind:
          default: drilldown
          title: Kind
          type: string
        label:
          title: Label
          type: string
        params:
          additionalProperties: true
          title: Params
          type: object
        targetTableId:
          anyOf:
            - type: string
            - type: 'null'
          title: Targettableid
      required:
        - id
        - label
      title: TableRowActionV1
      type: object
    TableRowV1:
      properties:
        actions:
          items:
            $ref: '#/components/schemas/TableRowActionV1'
          title: Actions
          type: array
        cells:
          additionalProperties: true
          title: Cells
          type: object
        id:
          title: Id
          type: string
      required:
        - id
      title: TableRowV1
      type: object
    TableV1Response:
      properties:
        columns:
          items:
            $ref: '#/components/schemas/TableColumnV1'
          title: Columns
          type: array
        commitId:
          anyOf:
            - type: string
            - type: 'null'
          title: Commitid
        rows:
          items:
            $ref: '#/components/schemas/TableRowV1'
          title: Rows
          type: array
        tableId:
          title: Tableid
          type: string
        title:
          title: Title
          type: string
      required:
        - tableId
        - title
      title: TableV1Response
      type: object
...
    ValidationError:
      properties:
        ctx:
          title: Context
          type: object
        input:
          title: Input
        loc:
          items:
            anyOf:
              - type: string
              - type: integer
          title: Location
          type: array
        msg:
          title: Message
          type: string
        type:
          title: Error Type
          type: string
      required:
        - loc
        - msg
        - type
      title: ValidationError
      type: object
...